ginkgo: Disable proximity sensor usage during doze
When the proximity sensor is in use, the smp2p_sleepstate
IRQ is fired multiple times a second, with each one
holding a 200ms wakelock. This is probably a bug in the
DSP firmware. To fix this, avoid using the proximity sensor
in doze mode, because sleep is preferred to turning off the
screen.

Change-Id: I5c54643b3d6d6ea4e5f847870a5266bb96342583
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Wrdn28 <ifloraxd@gmail.com>